requirements.txt
flask==3.0.2
flask-restx==1.3.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
